+++ ceammc/ext/src/synth/synth_metro.h
@@ -1,0 +1,905 @@
+/* ------------------------------------------------------------
+name: "synth.metro"
+Code generated with Faust 2.44.1 (https://faust.grame.fr)
+Compilation options: -a /Users/serge/work/music/pure-data/ceammc/faust/ceammc_dsp_ext.cpp -lang cpp -i -cn synth_metro -scn synth_metro_dsp -es 1 -mcd 16 -single -ftz 0
+------------------------------------------------------------ */
+
+#ifndef  __synth_metro_H__
+#define  __synth_metro_H__
+
+// FAUST Architecture File for ceammc::SoundExternal class
+#include <cmath>
+#include <cstdio>
+#include <cstdlib>
+#include <memory>
+#include <string>
+
+/************************** BEGIN synth_metro_dsp.h ********************************
+ FAUST Architecture File
+ Copyright (C) 2003-2022 GRAME, Centre National de Creation Musicale
+ ---------------------------------------------------------------------
+ This program is free software; you can redistribute it and/or modify
+ it under the terms of the GNU Lesser General Public License as published by
+ the Free Software Foundation; either version 2.1 of the License, or
+ (at your option) any later version.
+ 
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ GNU Lesser General Public License for more details.
+ 
+ You should have received a copy of the GNU Lesser General Public License
+ along with this program; if not, write to the Free Software
+ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
+ 
+ EXCEPTION : As a special exception, you may create a larger work
+ that contains this FAUST architecture section and distribute
+ that work under terms of your choice, so long as this FAUST
+ architecture section is not modified.
+ ************************************************************************/
+
+#ifndef __dsp__
+#define __dsp__
+
+#include <string>
+#include <vector>
+
+/************************************************************************
+ ************************************************************************
+    FAUST compiler
+    Copyright (C) 2003-2018 GRAME, Centre National de Creation Musicale
+    ---------------------------------------------------------------------
+    This program is free software; you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation; either version 2 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program; if not, write to the Free Software
+    Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
+ ************************************************************************
+ ************************************************************************/
+
+#ifndef __export__
+#define __export__
+
+#define FAUSTVERSION "2.44.1"
+
+// Use FAUST_API for code that is part of the external API but is also compiled in faust and libfaust
+// Use LIBFAUST_API for code that is compiled in faust and libfaust
+
+#ifdef _WIN32
+    #pragma warning (disable: 4251)
+    #ifdef FAUST_EXE
+        #define FAUST_API
+        #define LIBFAUST_API
+    #elif FAUST_LIB
+        #define FAUST_API __declspec(dllexport)
+        #define LIBFAUST_API __declspec(dllexport)
+    #else
+        #define FAUST_API
+        #define LIBFAUST_API 
+    #endif
+#else
+    #ifdef FAUST_EXE
+        #define FAUST_API
+        #define LIBFAUST_API
+    #else
+        #define FAUST_API __attribute__((visibility("default")))
+        #define LIBFAUST_API __attribute__((visibility("default")))
+    #endif
+#endif
+
+#endif
+
+#ifndef FAUSTFLOAT
+#define FAUSTFLOAT float
+#endif
+
+struct FAUST_API UI;
+struct FAUST_API Meta;
+
+/**
+ * DSP memory manager.
+ */
+
+struct FAUST_API dsp_memory_manager {
+    
+    virtual ~dsp_memory_manager() {}
+    
+    /**
+     * Inform the Memory Manager with the number of expected memory zones.
+     * @param count - the number of expected memory zones
+     */
+    virtual void begin(size_t count) {}
+    
+    /**
+     * Give the Memory Manager information on a given memory zone.
+     * @param size - the size in bytes of the memory zone
+     * @param reads - the number of Read access to the zone used to compute one frame
+     * @param writes - the number of Write access to the zone used to compute one frame
+     */
+    virtual void info(size_t size, size_t reads, size_t writes) {}
+    
+    /**
+     * Inform the Memory Manager that all memory zones have been described,
+     * to possibly start a 'compute the best allocation strategy' step.
+     */
+    virtual void end() {}
+    
+    /**
+     * Allocate a memory zone.
+     * @param size - the memory zone size in bytes
+     */
+    virtual void* allocate(size_t size) = 0;
+    
+    /**
+     * Destroy a memory zone.
+     * @param ptr - the memory zone pointer to be deallocated
+     */
+    virtual void destroy(void* ptr) = 0;
+    
+};
+
+/**
+* Signal processor definition.
+*/
+
+class FAUST_API synth_metro_dsp {
+
+    public:
+
+        synth_metro_dsp() {}
+        virtual ~synth_metro_dsp() {}
+
+        /* Return instance number of audio inputs */
+        virtual int getNumInputs() = 0;
+    
+        /* Return instance number of audio outputs */
+        virtual int getNumOutputs() = 0;
+    
+        /**
+         * Trigger the ui_interface parameter with instance specific calls
+         * to 'openTabBox', 'addButton', 'addVerticalSlider'... in order to build the UI.
+         *
+         * @param ui_interface - the user interface builder
+         */
+        virtual void buildUserInterface(UI* ui_interface) = 0;
+    
+        /* Return the sample rate currently used by the instance */
+        virtual int getSampleRate() = 0;
+    
+        /**
+         * Global init, calls the following methods:
+         * - static class 'classInit': static tables initialization
+         * - 'instanceInit': constants and instance state initialization
+         *
+         * @param sample_rate - the sampling rate in Hz
+         */
+        virtual void init(int sample_rate) = 0;
+
+        /**
+         * Init instance state
+         *
+         * @param sample_rate - the sampling rate in Hz
+         */
+        virtual void instanceInit(int sample_rate) = 0;
+    
+        /**
+         * Init instance constant state
+         *
+         * @param sample_rate - the sampling rate in Hz
+         */
+        virtual void instanceConstants(int sample_rate) = 0;
+    
+        /* Init default control parameters values */
+        virtual void instanceResetUserInterface() = 0;
+    
+        /* Init instance state (like delay lines...) but keep the control parameter values */
+        virtual void instanceClear() = 0;
+ 
+        /**
+         * Return a clone of the instance.
+         *
+         * @return a copy of the instance on success, otherwise a null pointer.
+         */
+        virtual synth_metro_dsp* clone() = 0;
+    
+        /**
+         * Trigger the Meta* parameter with instance specific calls to 'declare' (key, value) metadata.
+         *
+         * @param m - the Meta* meta user
+         */
+        virtual void metadata(Meta* m) = 0;
+    
+        /**
+         * DSP instance computation, to be called with successive in/out audio buffers.
+         *
+         * @param count - the number of frames to compute
+         * @param inputs - the input audio buffers as an array of non-interleaved FAUSTFLOAT samples (eiher float, double or quad)
+         * @param outputs - the output audio buffers as an array of non-interleaved FAUSTFLOAT samples (eiher float, double or quad)
+         *
+         */
+        virtual void compute(int count, FAUSTFLOAT** inputs, FAUSTFLOAT** outputs) = 0;
+    
+        /**
+         * DSP instance computation: alternative method to be used by subclasses.
+         *
+         * @param date_usec - the timestamp in microsec given by audio driver.
+         * @param count - the number of frames to compute
+         * @param inputs - the input audio buffers as an array of non-interleaved FAUSTFLOAT samples (either float, double or quad)
+         * @param outputs - the output audio buffers as an array of non-interleaved FAUSTFLOAT samples (either float, double or quad)
+         *
+         */
+        virtual void compute(double /*date_usec*/, int count, FAUSTFLOAT** inputs, FAUSTFLOAT** outputs) { compute(count, inputs, outputs); }
+       
+};
+
+/**
+ * Generic DSP decorator.
+ */
+
+class FAUST_API decorator_dsp : public synth_metro_dsp {
+
+    protected:
+
+        synth_metro_dsp* fDSP;
+
+    public:
+
+        decorator_dsp(synth_metro_dsp* synth_metro_dsp = nullptr):fDSP(synth_metro_dsp) {}
+        virtual ~decorator_dsp() { delete fDSP; }
+
+        virtual int getNumInputs() { return fDSP->getNumInputs(); }
+        virtual int getNumOutputs() { return fDSP->getNumOutputs(); }
+        virtual void buildUserInterface(UI* ui_interface) { fDSP->buildUserInterface(ui_interface); }
+        virtual int getSampleRate() { return fDSP->getSampleRate(); }
+        virtual void init(int sample_rate) { fDSP->init(sample_rate); }
+        virtual void instanceInit(int sample_rate) { fDSP->instanceInit(sample_rate); }
+        virtual void instanceConstants(int sample_rate) { fDSP->instanceConstants(sample_rate); }
+        virtual void instanceResetUserInterface() { fDSP->instanceResetUserInterface(); }
+        virtual void instanceClear() { fDSP->instanceClear(); }
+        virtual decorator_dsp* clone() { return new decorator_dsp(fDSP->clone()); }
+        virtual void metadata(Meta* m) { fDSP->metadata(m); }
+        // Beware: subclasses usually have to overload the two 'compute' methods
+        virtual void compute(int count, FAUSTFLOAT** inputs, FAUSTFLOAT** outputs) { fDSP->compute(count, inputs, outputs); }
+        virtual void compute(double date_usec, int count, FAUSTFLOAT** inputs, FAUSTFLOAT** outputs) { fDSP->compute(date_usec, count, inputs, outputs); }
+    
+};
+
+/**
+ * DSP factory class, used with LLVM and Interpreter backends
+ * to create DSP instances from a compiled DSP program.
+ */
+
+class FAUST_API dsp_factory {
+    
+    protected:
+    
+        // So that to force sub-classes to use deleteDSPFactory(dsp_factory* factory);
+        virtual ~dsp_factory() {}
+    
+    public:
+    
+        virtual std::string getName() = 0;
+        virtual std::string getSHAKey() = 0;
+        virtual std::string getDSPCode() = 0;
+        virtual std::string getCompileOptions() = 0;
+        virtual std::vector<std::string> getLibraryList() = 0;
+        virtual std::vector<std::string> getIncludePathnames() = 0;
+    
+        virtual synth_metro_dsp* createDSPInstance() = 0;
+    
+        virtual void setMemoryManager(dsp_memory_manager* manager) = 0;
+        virtual dsp_memory_manager* getMemoryManager() = 0;
+    
+};
+
+// Denormal handling
+
+#if defined (__SSE__)
+#include <xmmintrin.h>
+#endif
+
+class FAUST_API ScopedNoDenormals {
+    
+    private:
+    
+        intptr_t fpsr;
+        
+        void setFpStatusRegister(intptr_t fpsr_aux) noexcept
+        {
+        #if defined (__arm64__) || defined (__aarch64__)
+           asm volatile("msr fpcr, %0" : : "ri" (fpsr_aux));
+        #elif defined (__SSE__)
+            _mm_setcsr(static_cast<uint32_t>(fpsr_aux));
+        #endif
+        }
+        
+        void getFpStatusRegister() noexcept
+        {
+        #if defined (__arm64__) || defined (__aarch64__)
+            asm volatile("mrs %0, fpcr" : "=r" (fpsr));
+        #elif defined ( __SSE__)
+            fpsr = static_cast<intptr_t>(_mm_getcsr());
+        #endif
+        }
+    
+    public:
+    
+        ScopedNoDenormals() noexcept
+        {
+        #if defined (__arm64__) || defined (__aarch64__)
+            intptr_t mask = (1 << 24 /* FZ */);
+        #else
+            #if defined(__SSE__)
+            #if defined(__SSE2__)
+                intptr_t mask = 0x8040;
+            #else
+                intptr_t mask = 0x8000;
+            #endif
+            #else
+                intptr_t mask = 0x0000;
+            #endif
+        #endif
+            getFpStatusRegister();
+            setFpStatusRegister(fpsr | mask);
+        }
+        
+        ~ScopedNoDenormals() noexcept
+        {
+            setFpStatusRegister(fpsr);
+        }
+
+};
+
+#define AVOIDDENORMALS ScopedNoDenormals();
+
+#endif
+
+/************************** END synth_metro_dsp.h **************************/
+/************************** BEGIN UI.h *****************************
+ FAUST Architecture File
+ Copyright (C) 2003-2022 GRAME, Centre National de Creation Musicale
+ ---------------------------------------------------------------------
+ This program is free software; you can redistribute it and/or modify
+ it under the terms of the GNU Lesser General Public License as published by
+ the Free Software Foundation; either version 2.1 of the License, or
+ (at your option) any later version.
+ 
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ GNU Lesser General Public License for more details.
+ 
+ You should have received a copy of the GNU Lesser General Public License
+ along with this program; if not, write to the Free Software
+ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
+ 
+ EXCEPTION : As a special exception, you may create a larger work
+ that contains this FAUST architecture section and distribute
+ that work under terms of your choice, so long as this FAUST
+ architecture section is not modified.
+ ********************************************************************/
+
+#ifndef __UI_H__
+#define __UI_H__
+
+
+#ifndef FAUSTFLOAT
+#define FAUSTFLOAT float
+#endif
+
+/*******************************************************************************
+ * UI : Faust DSP User Interface
+ * User Interface as expected by the buildUserInterface() method of a DSP.
+ * This abstract class contains only the method that the Faust compiler can
+ * generate to describe a DSP user interface.
+ ******************************************************************************/
+
+struct Soundfile;
+
+template <typename REAL>
+struct FAUST_API UIReal {
+    
+    UIReal() {}
+    virtual ~UIReal() {}
+    
+    // -- widget's layouts
+    
+    virtual void openTabBox(const char* label) = 0;
+    virtual void openHorizontalBox(const char* label) = 0;
+    virtual void openVerticalBox(const char* label) = 0;
+    virtual void closeBox() = 0;
+    
+    // -- active widgets
+    
+    virtual void addButton(const char* label, REAL* zone) = 0;
+    virtual void addCheckButton(const char* label, REAL* zone) = 0;
+    virtual void addVerticalSlider(const char* label, REAL* zone, REAL init, REAL min, REAL max, REAL step) = 0;
+    virtual void addHorizontalSlider(const char* label, REAL* zone, REAL init, REAL min, REAL max, REAL step) = 0;
+    virtual void addNumEntry(const char* label, REAL* zone, REAL init, REAL min, REAL max, REAL step) = 0;
+    
+    // -- passive widgets
+    
+    virtual void addHorizontalBargraph(const char* label, REAL* zone, REAL min, REAL max) = 0;
+    virtual void addVerticalBargraph(const char* label, REAL* zone, REAL min, REAL max) = 0;
+    
+    // -- soundfiles
+    
+    virtual void addSoundfile(const char* label, const char* filename, Soundfile** sf_zone) = 0;
+    
+    // -- metadata declarations
+    
+    virtual void declare(REAL* zone, const char* key, const char* val) {}
+    
+    // To be used by LLVM client
+    virtual int sizeOfFAUSTFLOAT() { return sizeof(FAUSTFLOAT); }
+};
+
+struct FAUST_API UI : public UIReal<FAUSTFLOAT> {
+    UI() {}
+    virtual ~UI() {}
+};
+
+#endif
+/**************************  END  UI.h **************************/
+/************************** BEGIN meta.h *******************************
+ FAUST Architecture File
+ Copyright (C) 2003-2022 GRAME, Centre National de Creation Musicale
+ ---------------------------------------------------------------------
+ This program is free software; you can redistribute it and/or modify
+ it under the terms of the GNU Lesser General Public License as published by
+ the Free Software Foundation; either version 2.1 of the License, or
+ (at your option) any later version.
+ 
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ GNU Lesser General Public License for more details.
+ 
+ You should have received a copy of the GNU Lesser General Public License
+ along with this program; if not, write to the Free Software
+ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
+ 
+ EXCEPTION : As a special exception, you may create a larger work
+ that contains this FAUST architecture section and distribute
+ that work under terms of your choice, so long as this FAUST
+ architecture section is not modified.
+ ************************************************************************/
+
+#ifndef __meta__
+#define __meta__
+
+
+/**
+ The base class of Meta handler to be used in synth_metro_dsp::metadata(Meta* m) method to retrieve (key, value) metadata.
+ */
+struct FAUST_API Meta {
+    virtual ~Meta() {}
+    virtual void declare(const char* key, const char* value) = 0;
+};
+
+#endif
+/**************************  END  meta.h **************************/
+/************************** BEGIN misc.h *******************************
+FAUST Architecture File
+Copyright (C) 2003-2022 GRAME, Centre National de Creation Musicale
+---------------------------------------------------------------------
+This program is free software; you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation; either version 2.1 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License
+along with this program; if not, write to the Free Software
+Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
+
+EXCEPTION : As a special exception, you may create a larger work
+that contains this FAUST architecture section and distribute
+that work under terms of your choice, so long as this FAUST
+architecture section is not modified.
+***************************************************************************/
+
+#ifndef __misc__
+#define __misc__
+
+#include <algorithm>
+#include <map>
+#include <cstdlib>
+#include <string.h>
+#include <fstream>
+#include <string>
+
+
+struct MY_Meta : Meta, std::map<const char*, const char*>
+{
+    void declare(const char* key, const char* value) { (*this)[key] = value; }
+};
+
+static int lsr(int x, int n) { return int(((unsigned int)x) >> n); }
+
+static int int2pow2(int x) { int r = 0; while ((1<<r) < x) r++; return r; }
+
+static long lopt(char* argv[], const char* name, long def)
+{
+    for (int i = 0; argv[i]; i++) if (!strcmp(argv[i], name)) return std::atoi(argv[i+1]);
+    return def;
+}
+
+static long lopt1(int argc, char* argv[], const char* longname, const char* shortname, long def)
+{
+    for (int i = 2; i < argc; i++) {
+        if (strcmp(argv[i-1], shortname) == 0 || strcmp(argv[i-1], longname) == 0) {
+            return atoi(argv[i]);
+        }
+    }
+    return def;
+}
+
+static const char* lopts(char* argv[], const char* name, const char* def)
+{
+    for (int i = 0; argv[i]; i++) if (!strcmp(argv[i], name)) return argv[i+1];
+    return def;
+}
+
+static const char* lopts1(int argc, char* argv[], const char* longname, const char* shortname, const char* def)
+{
+    for (int i = 2; i < argc; i++) {
+        if (strcmp(argv[i-1], shortname) == 0 || strcmp(argv[i-1], longname) == 0) {
+            return argv[i];
+        }
+    }
+    return def;
+}
+
+static bool isopt(char* argv[], const char* name)
+{
+    for (int i = 0; argv[i]; i++) if (!strcmp(argv[i], name)) return true;
+    return false;
+}
+
+static std::string pathToContent(const std::string& path)
+{
+    std::ifstream file(path.c_str(), std::ifstream::binary);
+    
+    file.seekg(0, file.end);
+    int size = int(file.tellg());
+    file.seekg(0, file.beg);
+    
+    // And allocate buffer to that a single line can be read...
+    char* buffer = new char[size + 1];
+    file.read(buffer, size);
+    
+    // Terminate the string
+    buffer[size] = 0;
+    std::string result = buffer;
+    file.close();
+    delete [] buffer;
+    return result;
+}
+
+#endif
+
+/**************************  END  misc.h **************************/
+
+#include "ceammc_faust.h"
+
+using namespace ceammc::faust;
+
+#ifdef FAUST_MACRO
+// clang-format off
+// clang-format on
+#endif
+
+#define sym(name) xsym(name)
+#define xsym(name) #name
+
+// clang-format off
+#ifndef FAUST_MACRO
+struct synth_metro : public synth_metro_dsp {
+};
+#endif
+// clang-format on
+
+#ifdef FAUST_MACRO
+// clang-format off
+#ifndef FAUSTFLOAT
+#define FAUSTFLOAT float
+#endif 
+
+#include <algorithm>
+#include <cmath>
+#include <cstdint>
+#include <math.h>
+
+#ifndef FAUSTCLASS 
+#define FAUSTCLASS synth_metro
+#endif
+
+#ifdef __APPLE__ 
+#define exp10f __exp10f
+#define exp10 __exp10
+#endif
+
+#if defined(_WIN32)
+#define RESTRICT __restrict
+#else
+#define RESTRICT __restrict__
+#endif
+
+static float synth_metro_faustpower2_f(float value) {
+	return value * value;
+}
+
+class synth_metro : public synth_metro_dsp {
+	
+ private:
+	
+	int fSampleRate;
+	float fConst0;
+	FAUSTFLOAT fVslider0;
+	FAUSTFLOAT fVslider1;
+	FAUSTFLOAT fVslider2;
+	FAUSTFLOAT fButton0;
+	float fVec0[2];
+	float fRec1[3];
+	float fRec0[3];
+	FAUSTFLOAT fVslider3;
+	FAUSTFLOAT fVslider4;
+	FAUSTFLOAT fVslider5;
+	FAUSTFLOAT fButton1;
+	float fVec1[2];
+	float fRec3[3];
+	float fRec2[3];
+	FAUSTFLOAT fVslider6;
+	FAUSTFLOAT fVslider7;
+	FAUSTFLOAT fVslider8;
+	FAUSTFLOAT fButton2;
+	float fVec2[2];
+	float fRec5[3];
+	float fRec4[3];
+	
+ public:
+	
+	void metadata(Meta* m) { 
+		m->declare("basics.lib/name", "Faust Basic Element Library");
+		m->declare("basics.lib/version", "0.8");
+		m->declare("compile_options", "-a /Users/serge/work/music/pure-data/ceammc/faust/ceammc_dsp_ext.cpp -lang cpp -i -cn synth_metro -scn synth_metro_dsp -es 1 -mcd 16 -single -ftz 0");
+		m->declare("filename", "synth_metro.dsp");
+		m->declare("filters.lib/fir:author", "Julius O. Smith III");
+		m->declare("filters.lib/fir:copyright", "Copyright (C) 2003-2019 by Julius O. Smith III <jos@ccrma.stanford.edu>");
+		m->declare("filters.lib/fir:license", "MIT-style STK-4.3 license");
+		m->declare("filters.lib/iir:author", "Julius O. Smith III");
+		m->declare("filters.lib/iir:copyright", "Copyright (C) 2003-2019 by Julius O. Smith III <jos@ccrma.stanford.edu>");
+		m->declare("filters.lib/iir:license", "MIT-style STK-4.3 license");
+		m->declare("filters.lib/lowpass0_highpass1", "MIT-style STK-4.3 license");
+		m->declare("filters.lib/lowpass0_highpass1:author", "Julius O. Smith III");
+		m->declare("filters.lib/lowpass:author", "Julius O. Smith III");
+		m->declare("filters.lib/lowpass:copyright", "Copyright (C) 2003-2019 by Julius O. Smith III <jos@ccrma.stanford.edu>");
+		m->declare("filters.lib/lowpass:license", "MIT-style STK-4.3 license");
+		m->declare("filters.lib/name", "Faust Filters Library");
+		m->declare("filters.lib/resonbp:author", "Julius O. Smith III");
+		m->declare("filters.lib/resonbp:copyright", "Copyright (C) 2003-2019 by Julius O. Smith III <jos@ccrma.stanford.edu>");
+		m->declare("filters.lib/resonbp:license", "MIT-style STK-4.3 license");
+		m->declare("filters.lib/tf2:author", "Julius O. Smith III");
+		m->declare("filters.lib/tf2:copyright", "Copyright (C) 2003-2019 by Julius O. Smith III <jos@ccrma.stanford.edu>");
+		m->declare("filters.lib/tf2:license", "MIT-style STK-4.3 license");
+		m->declare("filters.lib/tf2s:author", "Julius O. Smith III");
+		m->declare("filters.lib/tf2s:copyright", "Copyright (C) 2003-2019 by Julius O. Smith III <jos@ccrma.stanford.edu>");
+		m->declare("filters.lib/tf2s:license", "MIT-style STK-4.3 license");
+		m->declare("filters.lib/version", "0.3");
+		m->declare("maths.lib/author", "GRAME");
+		m->declare("maths.lib/copyright", "GRAME");
+		m->declare("maths.lib/license", "LGPL with exception");
+		m->declare("maths.lib/name", "Faust Math Library");
+		m->declare("maths.lib/version", "2.5");
+		m->declare("name", "synth.metro");
+		m->declare("platform.lib/name", "Generic Platform Library");
+		m->declare("platform.lib/version", "0.2");
+	}
+
+	virtual int getNumInputs() {
+		return 0;
+	}
+	virtual int getNumOutputs() {
+		return 1;
+	}
+	
+	static void classInit(int sample_rate) {
+	}
+	
+	virtual void instanceConstants(int sample_rate) {
+		fSampleRate = sample_rate;
+		fConst0 = 3.14159274f / std::min<float>(192000.0f, std::max<float>(1.0f, float(fSampleRate)));
+	}
+	
+	virtual void instanceResetUserInterface() {
+		fVslider0 = FAUSTFLOAT(2000.0f);
+		fVslider1 = FAUSTFLOAT(3.0f);
+		fVslider2 = FAUSTFLOAT(30.0f);
+		fButton0 = FAUSTFLOAT(0.0f);
+		fVslider3 = FAUSTFLOAT(1000.0f);
+		fVslider4 = FAUSTFLOAT(0.0f);
+		fVslider5 = FAUSTFLOAT(10.0f);
+		fButton1 = FAUSTFLOAT(0.0f);
+		fVslider6 = FAUSTFLOAT(1500.0f);
+		fVslider7 = FAUSTFLOAT(3.0f);
+		fVslider8 = FAUSTFLOAT(300.0f);
+		fButton2 = FAUSTFLOAT(0.0f);
+	}
+	
+	virtual void instanceClear() {
+		for (int l0 = 0; l0 < 2; l0 = l0 + 1) {
+			fVec0[l0] = 0.0f;
+		}
+		for (int l1 = 0; l1 < 3; l1 = l1 + 1) {
+			fRec1[l1] = 0.0f;
+		}
+		for (int l2 = 0; l2 < 3; l2 = l2 + 1) {
+			fRec0[l2] = 0.0f;
+		}
+		for (int l3 = 0; l3 < 2; l3 = l3 + 1) {
+			fVec1[l3] = 0.0f;
+		}
+		for (int l4 = 0; l4 < 3; l4 = l4 + 1) {
+			fRec3[l4] = 0.0f;
+		}
+		for (int l5 = 0; l5 < 3; l5 = l5 + 1) {
+			fRec2[l5] = 0.0f;
+		}
+		for (int l6 = 0; l6 < 2; l6 = l6 + 1) {
+			fVec2[l6] = 0.0f;
+		}
+		for (int l7 = 0; l7 < 3; l7 = l7 + 1) {
+			fRec5[l7] = 0.0f;
+		}
+		for (int l8 = 0; l8 < 3; l8 = l8 + 1) {
+			fRec4[l8] = 0.0f;
+		}
+	}
+	
+	virtual void init(int sample_rate) {
+		classInit(sample_rate);
+		instanceInit(sample_rate);
+	}
+	virtual void instanceInit(int sample_rate) {
+		instanceConstants(sample_rate);
+		instanceResetUserInterface();
+		instanceClear();
+	}
+	
+	virtual synth_metro* clone() {
+		return new synth_metro();
+	}
+	
+	virtual int getSampleRate() {
+		return fSampleRate;
+	}
+	
+	virtual void buildUserInterface(UI* ui_interface) {
+		ui_interface->openVerticalBox("synth.metro");
+		ui_interface->addButton("b0", &fButton0);
+		ui_interface->addButton("b1", &fButton1);
+		ui_interface->addButton("b2", &fButton2);
+		ui_interface->declare(&fVslider0, "unit", "Hz");
+		ui_interface->addVerticalSlider("freq0", &fVslider0, FAUSTFLOAT(2000.0f), FAUSTFLOAT(100.0f), FAUSTFLOAT(10000.0f), FAUSTFLOAT(0.100000001f));
+		ui_interface->declare(&fVslider3, "unit", "Hz");
+		ui_interface->addVerticalSlider("freq1", &fVslider3, FAUSTFLOAT(1000.0f), FAUSTFLOAT(100.0f), FAUSTFLOAT(10000.0f), FAUSTFLOAT(0.100000001f));
+		ui_interface->declare(&fVslider6, "unit", "Hz");
+		ui_interface->addVerticalSlider("freq2", &fVslider6, FAUSTFLOAT(1500.0f), FAUSTFLOAT(100.0f), FAUSTFLOAT(10000.0f), FAUSTFLOAT(0.100000001f));
+		ui_interface->declare(&fVslider1, "unit", "db");
+		ui_interface->addVerticalSlider("gain0", &fVslider1, FAUSTFLOAT(3.0f), FAUSTFLOAT(-60.0f), FAUSTFLOAT(12.0f), FAUSTFLOAT(0.100000001f));
+		ui_interface->declare(&fVslider4, "unit", "db");
+		ui_interface->addVerticalSlider("gain1", &fVslider4, FAUSTFLOAT(0.0f), FAUSTFLOAT(-60.0f), FAUSTFLOAT(12.0f), FAUSTFLOAT(0.100000001f));
+		ui_interface->declare(&fVslider7, "unit", "db");
+		ui_interface->addVerticalSlider("gain2", &fVslider7, FAUSTFLOAT(3.0f), FAUSTFLOAT(-60.0f), FAUSTFLOAT(12.0f), FAUSTFLOAT(0.100000001f));
+		ui_interface->addVerticalSlider("q0", &fVslider2, FAUSTFLOAT(30.0f), FAUSTFLOAT(10.0f), FAUSTFLOAT(200.0f), FAUSTFLOAT(0.100000001f));
+		ui_interface->addVerticalSlider("q1", &fVslider5, FAUSTFLOAT(10.0f), FAUSTFLOAT(10.0f), FAUSTFLOAT(200.0f), FAUSTFLOAT(0.100000001f));
+		ui_interface->addVerticalSlider("q2", &fVslider8, FAUSTFLOAT(300.0f), FAUSTFLOAT(10.0f), FAUSTFLOAT(300.0f), FAUSTFLOAT(0.100000001f));
+		ui_interface->closeBox();
+	}
+	
+	virtual void compute(int count, FAUSTFLOAT** RESTRICT inputs, FAUSTFLOAT** RESTRICT outputs) {
+		FAUSTFLOAT* output0 = outputs[0];
+		float fSlow0 = std::tan(fConst0 * float(fVslider0));
+		float fSlow1 = 1.0f / fSlow0;
+		float fSlow2 = 1.0f / ((fSlow1 + 1.41421354f) / fSlow0 + 1.0f);
+		float fSlow3 = 1.0f / float(fVslider2);
+		float fSlow4 = (fSlow3 + fSlow1) / fSlow0 + 1.0f;
+		float fSlow5 = std::pow(10.0f, 0.0500000007f * (float(fVslider1) + 3.0f)) / (fSlow0 * fSlow4);
+		float fSlow6 = float(fButton0);
+		float fSlow7 = 1.0f / fSlow4;
+		float fSlow8 = (fSlow1 - fSlow3) / fSlow0 + 1.0f;
+		float fSlow9 = 2.0f * (1.0f - 1.0f / synth_metro_faustpower2_f(fSlow0));
+		float fSlow10 = 0.0f - fSlow5;
+		float fSlow11 = (fSlow1 + -1.41421354f) / fSlow0 + 1.0f;
+		float fSlow12 = std::tan(fConst0 * float(fVslider3));
+		float fSlow13 = 1.0f / fSlow12;
+		float fSlow14 = 1.0f / ((fSlow13 + 1.41421354f) / fSlow12 + 1.0f);
+		float fSlow15 = 1.0f / float(fVslider5);
+		float fSlow16 = (fSlow15 + fSlow13) / fSlow12 + 1.0f;
+		float fSlow17 = std::pow(10.0f, 0.0500000007f * (float(fVslider4) + 3.0f)) / (fSlow12 * fSlow16);
+		float fSlow18 = float(fButton1);
+		float fSlow19 = 1.0f / fSlow16;
+		float fSlow20 = (fSlow13 - fSlow15) / fSlow12 + 1.0f;
+		float fSlow21 = 2.0f * (1.0f - 1.0f / synth_metro_faustpower2_f(fSlow12));
+		float fSlow22 = 0.0f - fSlow17;
+		float fSlow23 = (fSlow13 + -1.41421354f) / fSlow12 + 1.0f;
+		float fSlow24 = std::tan(fConst0 * float(fVslider6));
+		float fSlow25 = 1.0f / fSlow24;
+		float fSlow26 = 1.0f / ((fSlow25 + 1.41421354f) / fSlow24 + 1.0f);
+		float fSlow27 = 1.0f / float(fVslider8);
+		float fSlow28 = (fSlow27 + fSlow25) / fSlow24 + 1.0f;
+		float fSlow29 = std::pow(10.0f, 0.0500000007f * (float(fVslider7) + 3.0f)) / (fSlow24 * fSlow28);
+		float fSlow30 = float(fButton2);
+		float fSlow31 = 1.0f / fSlow28;
+		float fSlow32 = (fSlow25 - fSlow27) / fSlow24 + 1.0f;
+		float fSlow33 = 2.0f * (1.0f - 1.0f / synth_metro_faustpower2_f(fSlow24));
+		float fSlow34 = 0.0f - fSlow29;
+		float fSlow35 = (fSlow25 + -1.41421354f) / fSlow24 + 1.0f;
+		for (int i0 = 0; i0 < count; i0 = i0 + 1) {
+			fVec0[0] = fSlow6;
+			float fTemp0 = fSlow6 - fVec0[1];
+			fRec1[0] = fTemp0 * float(fTemp0 > 0.0f) - fSlow7 * (fSlow8 * fRec1[2] + fSlow9 * fRec1[1]);
+			fRec0[0] = (fSlow5 * fRec1[0] + fSlow10 * fRec1[2]) - fSlow2 * (fSlow11 * fRec0[2] + fSlow9 * fRec0[1]);
+			fVec1[0] = fSlow18;
+			float fTemp1 = fSlow18 - fVec1[1];
+			fRec3[0] = fTemp1 * float(fTemp1 > 0.0f) - fSlow19 * (fSlow20 * fRec3[2] + fSlow21 * fRec3[1]);
+			fRec2[0] = (fSlow17 * fRec3[0] + fSlow22 * fRec3[2]) - fSlow14 * (fSlow23 * fRec2[2] + fSlow21 * fRec2[1]);
+			fVec2[0] = fSlow30;
+			float fTemp2 = fSlow30 - fVec2[1];
+			fRec5[0] = fTemp2 * float(fTemp2 > 0.0f) - fSlow31 * (fSlow32 * fRec5[2] + fSlow33 * fRec5[1]);
+			fRec4[0] = (fSlow29 * fRec5[0] + fSlow34 * fRec5[2]) - fSlow26 * (fSlow35 * fRec4[2] + fSlow33 * fRec4[1]);
+			output0[i0] = FAUSTFLOAT(fSlow2 * (fRec0[2] + fRec0[0] + 2.0f * fRec0[1]) + fSlow14 * (fRec2[2] + fRec2[0] + 2.0f * fRec2[1]) + fSlow26 * (fRec4[2] + fRec4[0] + 2.0f * fRec4[1]));
+			fVec0[1] = fVec0[0];
+			fRec1[2] = fRec1[1];
+			fRec1[1] = fRec1[0];
+			fRec0[2] = fRec0[1];
+			fRec0[1] = fRec0[0];
+			fVec1[1] = fVec1[0];
+			fRec3[2] = fRec3[1];
+			fRec3[1] = fRec3[0];
+			fRec2[2] = fRec2[1];
+			fRec2[1] = fRec2[0];
+			fVec2[1] = fVec2[0];
+			fRec5[2] = fRec5[1];
+			fRec5[1] = fRec5[0];
+			fRec4[2] = fRec4[1];
+			fRec4[1] = fRec4[0];
+		}
+	}
+
+};
+// clang-format on
+#endif
+
+    template <class T>
+    struct _synth_metro_UI : public UI {
+    static std::string name;
+};
+
+template <class T>
+std::string _synth_metro_UI<T>::name(sym(synth_metro));
+
+typedef _synth_metro_UI<synth_metro> synth_metro_UI;
+
+class faust_synth_metro_tilde : public FaustExternal<synth_metro, synth_metro_UI> {
+public:
+    faust_synth_metro_tilde(const ceammc::PdArgs& args)
+        : FaustExternal(args)
+    {
+    }
+};
+
+#endif